requirements.txt
Django==3.2.4
django-rest-framework
djangorestframework
markdown
django-filter
psycopg2-binary
PyJWT
djoser
djangorestframework_simplejwt~=4.3.0
python-dotenv
drf-spectacular
drf-psq
gunicorn==20.1.0
Pillow